CRLF -> LF
GDI why does this always happen :/ I need to stop using Windows tools
when I'm lazy and don't develop in my VM lol. Need to be using at least
Sublime or even emacs/vim so that it's LF always.

I was going to do a git filter branch but whatever this is easier.