Improve accessibility of HTML elements
Using the newer main, article, aside, footer tags. I feel like this also
keeps the code cleaner and more obvious for what's going on. Will
continue to improve this as I said. But will also need to test on older
systems and text-based browsers to ensure compatability (because of
course I care).

References:
- https://developer.mozilla.org/en-US/docs/Learn/Accessibility/HTML
- MDN's sample page: https://mdn.github.io/learning-area/html/introduction-to-html/document_and_website_structure/