Add image figure caption support + captions + some image licence
Some images I will license under CC as I decide. I figured this would
probably be the most appropriate way to go about it, but this then
prompted caption support for images. Noticed that figures might be the
way to go.

Turns out Hugo has a great Markdown templating system that allows custom
templates for elements such as images! Not sure if Jekyll had this but
definitely cool to see here. Thanks Sebastian De Deyne for this post
https://sebastiandedeyne.com/captioned-images-with-markdown-render-hooks-in-hugo/,
essentially showed me an awesome, neat way to handle captions with the
"title" field!