Switch to 'pretty' urls :<
This isn't a style change I personally wanted, but seems like Cloudflare
Pages won't allow it and it's causing 308 redirects all the time when
hitting the .html links and redirecting to the non-.html link. :( Might
follow this one up on their forums but just sounds like they don't care.