Rewrite life timeline generator to be statically generated
Have wanted to do this for a while, given the AJAX getting of config and
the timeline content itself has always felt pretty redundant to me,
given that we are generating at runtime anyway.

Only the scroll is left in JS :) So even with no JS this is fully
functional just without the QoL mouse drag scroll.

Now just need to clean up the styling and the format of the events,
given it's a little hacky right now with the `dateText` field to work
around having to do even more complicated logic in Go Templates.