Use upgrade-insecure-requests CSP
I wanted to upgrade HTTP requests to HTTPS for my site, but only when it
makes sense to do so - i.e. older text based / simple browsers should
not bother. Also need a way to do it without any powerful UA inspection
or other server-side techniques since I'm using static GitHub hosting.
This seems perfect, since older browsers are probably too old to respect
this meta! Ideally it'd be sent as a header, but it's nice this works
too.